splitted components into separate files
